fix(CICD): Modify Jenkinsfile (#142)
* fix(CICD): Modify Jenkinsfile

* covert table

* format

+++ Jenkinsfile
@@ -116,50 +116,65 @@ podTemplate(
                 }
 
                 stage('Run e2e test') {
-                    sh('''
-                        set -e
-                        cd ${WORKDIR}
-                        # get host ip
-                        HOST_IP=$(ifconfig eth0 | grep 'inet addr:'| grep -v '127.0.0.1' | cut -d: -f2 | awk '{ print $1}')
-                        export CYCLONE_SERVER=http://${HOST_IP}:7099
-                        export LOG_SERVER=ws://${HOST_IP}:8000/ws
+                    if (params.integration) {
+                        echo "exec integration"
+                        sh('''
+                            set -e
+                            cd ${WORKDIR}
+                            # get host ip
+                            HOST_IP=$(ifconfig eth0 | grep 'inet addr:'| grep -v '127.0.0.1' | cut -d: -f2 | awk '{ print $1}')
+                            export CYCLONE_SERVER=http://${HOST_IP}:7099
+                            export LOG_SERVER=ws://${HOST_IP}:8000/ws
                         
-                        # kill cyclone server process left by last build
-                        cyclone_pid=$(ps -ef | grep cyclone-server | grep -v "grep" | awk '{print $1}')
-                        if [[ -n "${cyclone_pid}" ]]; then
-                            kill -9 ${cyclone_pid}
-                        fi
+                            # kill cyclone server process left by last build
+                            cyclone_pid=$(ps -ef | grep cyclone-server | grep -v "grep" | awk '{print $1}')
+                            if [[ -n "${cyclone_pid}" ]]; then
+                                kill -9 ${cyclone_pid}
+                            fi
 
-                        echo "start server"
-                        ./cyclone-server --cloud-auto-discovery=false --log-force-color=true &
+                            echo "start server"
+                            ./cyclone-server --cloud-auto-discovery=false --log-force-color=true &
 
-                        echo "testing ..."
-                        # go test compile
-                        go test -i ./tests/...
+                            echo "testing ..."
+                            # go test compile
+                            go test -i ./tests/...
                         
-                        # go test
-                        go test -v ./tests/service 
-                        go test -v ./tests/version 
-                        go test -v ./tests/yaml
-                    ''')
+                            # go test
+                            go test -v ./tests/service 
+                            go test -v ./tests/version 
+                            go test -v ./tests/yaml
+                        ''')
+                    } else {
+                        echo "skip integration"
+                    }
                 }
             }
 
             stage("Build image and publish") {
-                sh "docker build -t ${server_tag} -f Dockerfile.server ."
-                sh "docker build -t ${worker_tag} -f Dockerfile.worker ."
+                if (params.publish) {
+                   echo "exec publish"
+                   sh "docker build -t ${server_tag} -f Dockerfile.server ."
+                   sh "docker build -t ${worker_tag} -f Dockerfile.worker ."
 
-                docker.withRegistry("https://${registry}", "cargo-private-admin") {
-                    docker.image(server_tag).push()
-                    docker.image(worker_tag).push()
+                   docker.withRegistry("https://${registry}", "cargo-private-admin") {
+                       docker.image(server_tag).push()
+                       docker.image(worker_tag).push()
+	               } 
+                } else {
+                    echo "skip publish"	
                 }
             }
         }
 
         stage("deploy") {
-            sh("""
-                kubectl --namespace cyclone get deploy circle-server-v0.0.1 -o yaml | sed 's/cyclone-server:.*\$/cyclone-server:${env.BUILD_NUMBER}/; s/cyclone-worker:.*\$/cyclone-worker:${env.BUILD_NUMBER}/' | kubectl --namespace cyclone replace -f -
-            """)
+            if (params.deploy) {
+                echo "exec deploy"
+                sh("""
+                    kubectl --namespace cyclone get deploy circle-server-v0.0.1 -o yaml | sed 's/cyclone-server:.*\$/cyclone-server:${env.BUILD_NUMBER}/; s/cyclone-worker:.*\$/cyclone-worker:${env.BUILD_NUMBER}/' | kubectl --namespace cyclone replace -f -
+                """)
+            } else {
+                echo "skip deploy"	
+            }
         }
     }
 }
